Project structure refactored
