+++ requirements.txt
@@ -1,14 +1,14 @@
 Flask==1.1.2
 flask_restful==0.3.8
 sqlalchemy==1.4.0
-setuptools~=54.1.2
+setuptools==54.1.2
 requests==2.25.1
 redis==3.5.3
 psycopg2-binary==2.8.6
 marshmallow==3.10
 requests==2.25.1
-selenium~=3.141.0
-beautifulsoup4~=4.9.3
-alembic~=1.5.7
-click~=7.1.2
+selenium==3.141.0
+beautifulsoup4==4.9.3
+alembic==1.5.7
+click==7.1.2
 celery==5.0.5
