+++ service_api/celery_tasks/updaters.py
@@ -10,13 +10,13 @@ import requests
 from sqlalchemy.engine.row import Row
 
 from service_api import session_scope, models, schemas, LOGGER
-from service_api.grabbing_api.utils.limitation import DomriaLimitationSystem
-from service_api.errors import InternalServerErrorException
-from service_api.exceptions import ResponseNotOkException, MetaDataError
-from service_api.grabbing_api.constants import PATH_TO_METADATA
-from service_api.grabbing_api.utils.grabbing_utils import open_metadata, load_data
-from service_api.grabbing_api.utils.services_convertors import DomRiaOutputConverter
-from service_api.models import Realty, RealtyDetails
+from ..services.domria.limitation import DomriaLimitationSystem
+from ..errors import InternalServerErrorException
+from ..exceptions import ResponseNotOkException, MetaDataError
+from ..constants import PATH_TO_METADATA
+from ..utils import open_metadata, load_data
+from ..services.domria.convertors import DomRiaOutputConverter
+from ..models import Realty, RealtyDetails
 
 
 class AbstractUpdater(ABC):
