+++ service_api/grabbing_api/__init__.py
@@ -1,4 +1,3 @@
 """
 Main module of grabbing service contains routes and utils for filling db and getting data
 """
-from . import constants, resources
