+++ service_api/grabbing_api/resources.py
@@ -6,10 +6,10 @@ from flask import request
 from flask_restful import Resource
 
 from service_api import api_
-from service_api.constants import URLS
-from service_api.errors import InternalServerErrorException
-from service_api.exceptions import MetaDataError
-from .utils.db import RealtyFetcher
+from ..constants import URLS
+from ..errors import InternalServerErrorException
+from ..exceptions import MetaDataError
+from ..utils.db import RealtyFetcher
 
 
 class LatestDataResource(Resource):
