+++ service_api/services/domria/__init__.py
@@ -1,0 +1,3 @@
+"""
+Domria logic
+"""
