+++ service_api/services/domria/limitation.py
@@ -1,18 +1,19 @@
 """
-Limitation system module
+Domria limitation system
 """
+
+
 from datetime import datetime, timedelta
 from hashlib import sha256
+from urllib.parse import parse_qs, urlencode, urlparse, urlunparse
 
-from urllib.parse import urlencode, urlparse, urlunparse, parse_qs
 from marshmallow.exceptions import ValidationError
-
 from service_api import LOGGER, session_scope
-from service_api.exceptions import LimitBoundError
-from service_api.grabbing_api.constants import DOMRIA_TOKENS_LIST, PATH_TO_METADATA, PATH_PARSER_METADATA
-from service_api.models import RequestsHistory
-from service_api.schemas import RequestsHistorySchema
-from .grabbing_utils import open_metadata
+
+from ...constants import DOMRIA_TOKENS_LIST
+from ...exceptions import LimitBoundError
+from ...models import RequestsHistory
+from ...schemas import RequestsHistorySchema
 
 
 class DomriaLimitationSystem:
@@ -68,34 +69,3 @@ class DomriaLimitationSystem:
             cls.TOKENS.append(cls.TOKENS.pop(0))
 
         raise LimitBoundError("DOMRIA limit reached! Try later!")
-
-
-class Singleton(type):
-    """
-    Implementation of sibgletopn pattern
-    """
-
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super().__call__(*args, **kwargs)
-        return cls._instances[cls]
-
-
-class LimitationSystem(metaclass=Singleton):
-    """
-    Control number of request to services
-    """
-    SERVICES = {"DOMRIA API": DomriaLimitationSystem}
-
-    def __init__(self) -> None:
-        self.metadata = open_metadata(PATH_TO_METADATA) | open_metadata(PATH_PARSER_METADATA)
-
-    def mark_token_after_request(self, url: str):
-        """
-        Check if such url must be supervised and if so then mark it as used
-        """
-        for service in self.metadata.values():
-            if url.startswith(service["base_url"]):
-                self.SERVICES[service["name"]].mark_token_after_requset(url)
