+++ service_api/services/olx/__init__.py
@@ -1,0 +1,3 @@
+"""
+Olx logic
+"""
